Move models into centralized file, set relationships, and simplify column declarations
